Extended versions to longs to address issues like https://help.octopus.com/t/cant-deploy-because-of-package-versioning-validation-error/26242

+++ source/Octopus.Versioning/IVersion.cs
@@ -29,10 +29,10 @@ namespace Octopus.Versioning
     /// </summary>
     public interface IVersion : IComparable
     {
-        int Major { get; }
-        int Minor { get; }
-        int Patch { get; }
-        int Revision { get; }
+        long Major { get; }
+        long Minor { get; }
+        long Patch { get; }
+        long Revision { get; }
         bool IsPrerelease { get; }
         IEnumerable<string> ReleaseLabels { get; }
         string? Metadata { get; }
